fix: 리뷰 피드백 반영
- 로그인 시, 가입 이력 체크 로직 변경 (email -> provider + providerId)
- UserProfileRepository ID 타입 오타 수정 (Integer -> Long)
- Refresh 토큰 생성 로직 변경 (JWT -> UUID)

+++ backend/src/main/java/com/ll/readycode/global/common/auth/user/CustomUserDetailsService.java
@@ -22,7 +22,7 @@ public class CustomUserDetailsService implements UserDetailsService {
 
     UserProfile userProfile =
         userProfileRepository
-            .findById(Integer.parseInt(userId))
+            .findById(Long.parseLong(userId))
             .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
 
     return new CustomUserDetails(userProfile);
